+++ src/handler.js
@@ -14,9 +14,32 @@ export async function processImage ({ Records: records }, context, callback) {
   return callback(null)
 }
 
-export async function metadata ({ Records: records }, context, callback) {
+export async function processItemMetadata ({ Records: records }, context, callback) {
   try {
     await Promise.all(records.map(processItem))
+
+    /*
+    let middlewareResult = true
+
+    if (metadata.rekognition) {
+      // @TODO
+      imageMetadata.rekognition = {}
+    }
+
+    if (metadata.middleware) {
+      middlewareResult = metadata.middleware(imageMetadata)
+    }
+
+    if (middlewareResult) {
+
+      if (metadata.saveJson) {
+        await upload(middlewareResult, {
+          ContentType: 'application/json',
+          Key: `${imageObjectKey}.json`,
+        })
+      }
+    }
+*/
   } catch (error) {
     console.error(error)
     return callback(error, { records })