+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - migrated from webpack 1 to webpack 3.3.0 and serverless-webpack 2.0
 - upgraded dependencies:  serverless@1.18.0, ava@0.21.0
-- **BREAKING:** converted the user `/config.json` file to JS format
+- **BREAKING:** converted the user `/config.json` file to JS format `/config.js` (#41)
 
 
 ## [0.11.0] - 2017-07-23